Move scenario construction (#1815)
