+++ smarts/diagnostic/run.py
@@ -34,9 +34,9 @@ import psutil
 from mdutils.mdutils import MdUtils
 
 import smarts
-from cli.studio import build_scenarios
 from smarts.core.scenario import Scenario
 from smarts.core.utils.math import welford
+from smarts.sstudio.scenario_construction import build_scenarios
 
 _SEED = 42
 _MAX_REPLAY_EPISODE_STEPS = 100
@@ -47,11 +47,7 @@ logger.setLevel(logging.INFO)
 
 
 def _compute(scenario_dir, ep_per_scenario=10, max_episode_steps=_MAX_EPISODE_STEPS):
-    build_scenarios(
-        clean=False,
-        scenarios=scenario_dir,
-        seed=_SEED,
-    )
+    build_scenarios(scenarios=scenario_dir, seed=_SEED)
     env = gym.make(
         "smarts.env:hiway-v0",
         scenarios=scenario_dir,
